feat: mobile accordion only migration
