+++ src/components/mobileOnlyAccordion/index.tsx
@@ -1,17 +1,17 @@
 import React, { FC, PropsWithChildren } from 'react';
-import { AccordionProps } from '@chakra-ui/react';
+import { Accordion as ChakraAccordion } from '@chakra-ui/react';
 
-import Accordion from '../accordion';
+import { Accordion } from '../accordion';
 
 export type Props = {
   variant?: 'light' | 'dark';
-} & Pick<AccordionProps, 'allowMultiple' | 'allowToggle'>;
+} & Pick<ChakraAccordion.RootProps, 'multiple' | 'collapsible'>;
 
 const MobileOnlyAccordion: FC<PropsWithChildren<Props>> = (props) => {
   const { children, ...restProps } = props;
 
   return (
-    <Accordion allowMultiple {...restProps}>
+    <Accordion multiple {...restProps}>
       {children}
     </Accordion>
   );
